% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ambiguity}
\alias{ambiguity}
\title{Data on party policy ambiguity}
\format{
An object of class \code{data.frame} with 321 rows and 4 columns.
}
\source{
\doi{https://doi.org/10.7910/DVN/GK9VCH}
}
\usage{
data(ambiguity)
}
\description{
Data from Eurobarometer surveys and political texts on voter 
and party positions in 14 Western European Democracies.
}
\examples{
data(ambiguity)
}
\references{
Braeuninger, Thomas; Giger, Nathalie, 2016, "Replication Data for: 
Strategic ambiguity of party positions in multiparty competition", 
\doi{https://doi.org/10.7910/DVN/GK9VCH}, Harvard Dataverse, V1.
}
\keyword{datasets}
