% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfun.R
\name{coef.qmix}
\alias{coef.qmix}
\title{Extract qmix Coefficients}
\usage{
\method{coef}{qmix}(object, ...)
}
\arguments{
\item{object}{A \code{qmix} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A table of coefficients with their corresponding lower and upper bounds.
}
\description{
Create a table of coefficient results from a \code{qmix} object.
}
