% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_fun.R
\name{dald}
\alias{dald}
\title{Probability density function of ALD}
\usage{
dald(x, mu, p, sigma)
}
\arguments{
\item{x}{Random variable.}

\item{mu}{Position parameter.}

\item{p}{Quantile.}

\item{sigma}{Scale parameter.}
}
\value{
probability density of \code{x}.
}
\description{
Probability density function of ALD
}
