% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flood}
\alias{flood}
\title{Flood data}
\format{
An object of class \code{data.frame} with 1662 rows and 5 columns.
}
\source{
\doi{https://doi.org/10.7910/DVN/W8US2Q}
}
\usage{
data(flood)
}
\description{
Recorded individual flood events from 1980 to 2008 in the world.
}
\examples{
data(flood)
}
\references{
Neumayer, Eric, Pluemper, Thomas, & Barthel, Fabian. 2014. 
The Political Economy of Natural Disaster Damage. Global Environmental Change, 24, 8-19.
}
\keyword{datasets}
