% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_fun.R
\name{gendata_qmix}
\alias{gendata_qmix}
\title{Generate synthetic data using finite mixture of quantiles}
\usage{
gendata_qmix(
  k = 2,
  n = 50,
  beta = c(-10, 10),
  sigma = c(1, 1),
  quantile = c(0.1, 0.9),
  seed = NULL
)
}
\arguments{
\item{k}{Number of mixture components. The default is 2.}

\item{n}{Number of observations for each component. The default is 50.}

\item{beta}{Coefficients.}

\item{sigma}{Scale parameters. The default is 1.}

\item{quantile}{Quantiles for the respective mixture components. The default values are 0.1 and 0.9 for the two-component mixture.}

\item{seed}{Seed for the random number generator.}
}
\value{
A data frame with outcomes Y and a covariate X.
}
\description{
\code{gendata_qmix} generates synthetic data using finite mixture of quantiles based on asymmetric Laplace distributions.
}
