% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_fun.R
\name{pald}
\alias{pald}
\title{Cumulative density function of ALD}
\usage{
pald(x, mu, p, sigma)
}
\arguments{
\item{x}{Random variable.}

\item{mu}{Position parameter.}

\item{p}{Quantile.}

\item{sigma}{Scale parameter.}
}
\value{
cumulative probability density of \code{x}.
}
\description{
Cumulative density function of ALD
}
