% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfun.R
\name{plot_coef.qmix}
\alias{plot_coef.qmix}
\title{Make coefficient plots for qmix}
\usage{
plot_coef.qmix(object, ...)
}
\arguments{
\item{object}{A \code{qmix} object.}

\item{...}{Additional parameters to be passed.}
}
\description{
Plot traceplots from a \code{qmix} object.
}
