% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfun.R
\name{plot_effect.qmix}
\alias{plot_effect.qmix}
\title{Plots of substantive effects for qmix}
\usage{
plot_effect.qmix(
  object,
  var = NULL,
  new_data = NULL,
  sim_var = NULL,
  xlab = "Variable of interest",
  ylab = "Substantive effects on outcomes",
  ...
)
}
\arguments{
\item{object}{A \code{qmix} object from running the main function \code{\link{qmix}}.}

\item{var}{Variable whose effect is to be ploted.}

\item{new_data}{New simulated data to be ploted.}

\item{sim_var}{A numeric vector of the simulated variable.}

\item{xlab}{Label of x-axis.}

\item{ylab}{Label of y-axis.}

\item{...}{Additional parameters to be passed to \code{\link[rstan]{traceplot}}.}
}
\value{
None.
}
\description{
\code{plot_effect.qmix} is used to produce plots of substantive effects from a \code{qmix} object from the main function \code{\link{qmix}}.
}
