% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfun.R
\name{plot_trace.qmix}
\alias{plot_trace.qmix}
\title{Make traceplots for qmix}
\usage{
plot_trace.qmix(object, ...)
}
\arguments{
\item{object}{A \code{qmix} object.}

\item{...}{Additional parameters to be passed.}
}
\description{
Plot traceplots from a \code{qmix} object.
}
