% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfun.R
\name{print_coef.qmix}
\alias{print_coef.qmix}
\title{Print qmix coefficients}
\usage{
print_coef.qmix(object, digits = 3)
}
\arguments{
\item{object}{A \code{qmix} object.}

\item{digits}{Number of digits to display.}
}
\description{
Print qmix coefficients
}
