% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfun.R
\name{print_mcmc.qmix}
\alias{print_mcmc.qmix}
\title{Print the mcmc results from a qmix object}
\usage{
print_mcmc.qmix(object, ...)
}
\arguments{
\item{object}{A \code{qmix} object.}

\item{...}{Additional arguments to be passed.}
}
\description{
This prints a number of diagnostics about the results of a \code{qmix} objects
}
