% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfun.R
\name{print_text.qmix}
\alias{print_text.qmix}
\title{Print the main results from a \code{qmix} object.}
\usage{
print_text.qmix(object, digits = 3)
}
\arguments{
\item{object}{A \code{qmix} object.}

\item{digits}{Number of digits to display.}
}
\value{
None.
}
\description{
Print the main results from a \code{qmix} object.
}
