% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_fun.R
\name{qald}
\alias{qald}
\title{Quantile function of ALD}
\usage{
qald(y, mu, p, sigma)
}
\arguments{
\item{y}{quantile value.}

\item{mu}{Position parameter.}

\item{p}{Quantile.}

\item{sigma}{Scale parameter.}
}
\value{
quantile value.
}
\description{
Quantile function of ALD
}
