% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmix-package.R
\docType{package}
\name{qmix-package}
\alias{qmix-package}
\title{'qmix': A R package for estimating finite quantile mixture models}
\description{
This package provides easy-to-use functionalities to estimate finite quantile mixture models with either fixed- or random-quantile specification.
}
\references{
Stan Development Team (2019). RStan: the R interface to Stan. R package version 2.19.2. https://mc-stan.org
}
