% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmix.R
\name{qmix}
\alias{qmix}
\title{Fitting conditional binary quantile models}
\usage{
qmix(formula, data, nmix = 3, design = "fixed", q = NULL,
  nsim = 1000, burnin = NULL, thin = 1, CIsize = 0.95,
  nchain = 1, seeds = 12345, offset = 1e-20, inverse_distr = FALSE)
}
\arguments{
\item{formula}{An object of class "formula" (or one that can be
coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{A data frame containing the variables in the model.}

\item{nmix}{The number of mixture components.}

\item{design}{Quantile specification. Options include "fixed" and "random". The default choice is "fixed" which requires quantile inputs from the user.}

\item{q}{The quantile value.}

\item{nsim}{The number of iterations.}

\item{burnin}{The number of burnin iterations.}

\item{thin}{Thinning parameter.}

\item{CIsize}{The size of posterior confidence interval.}

\item{nchain}{The number of parallel chains.}

\item{seeds}{Random seeds to replicate the results.}

\item{offset}{Offset values to enhance sampling stability. The default value is 1e-20.}

\item{inverse_distr}{If FALSE, the ALD will not be reversed. The default is FALSE.}
}
\value{
A \code{qmix} object. An object of class \code{qmix} contains the following elements

  \describe{

  \item{\code{Call}}{The matched call.}
  \item{\code{formula}}{Symbolic representation of the model.}

}
}
\description{
The main function for running the finite quantile mixture model. The function returns a \code{qmix} object that can be further investigated using standard functions such as \code{plot}, \code{print}, and \code{coef}. The model can be passed using a \code{formula} as in \code{lm()}. Convergence diagnotics can be performed using either \code{print(object, "mcmc")} or \code{plot(object, "mcmc")}.
}
