% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_fun.R
\name{inverse}
\alias{inverse}
\title{Inverse function}
\usage{
inverse(f, mu, p, sigma, lower = -10000, upper = 10000)
}
\arguments{
\item{f}{pald function}

\item{mu}{Position parameter.}

\item{p}{Quantile.}

\item{sigma}{Scale parameter.}

\item{lower}{Lower bound.}

\item{upper}{Upper bound.}
}
\value{
inversed pald
}
\description{
\code{inverse} generates inverse function of any given function.
}
