% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfun.R
\name{print_coef.qmix}
\alias{print_coef.qmix}
\title{Print coefficients of a \code{qmix} object}
\usage{
print_coef.qmix(object, digits = 3)
}
\arguments{
\item{object}{A \code{qmix} object.}

\item{digits}{Number of digits to display.}
}
\value{
None.
}
\description{
\code{print_coef.qmix} prints out coefficients from a \code{qmix} object from running the main function \code{\link{qmix}}.
}
