% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmix-package.R
\docType{package}
\name{qmix-package}
\alias{qmix-package}
\title{'qmix': A R Package for Estimating Finite Quantile Mixture Models}
\description{
This package provides easy-to-use functionalities to estimate finite quantile mixture models with either fixed- or random-quantile specification. The estimation is implemented using MCMC methods available in \code{rstan}.
}
\references{
Lu, Xiao (2019). Beyond the Average: Conditional Hypothesis Testing with Quantile Mixture.

Stan Development Team (2019). RStan: the R interface to Stan. R package version 2.19.2. https://mc-stan.org
}
