% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_fun.R
\name{rald}
\alias{rald}
\title{Random number generator of asymmetric Laplace distributions}
\usage{
rald(n, mu, p, sigma)
}
\arguments{
\item{n}{Number of random numbers to be generated.}

\item{mu}{Position parameter.}

\item{p}{Quantile.}

\item{sigma}{Scale parameter.}
}
\value{
random numbers.
}
\description{
\code{rald} generates random numbers from asymmetric Laplace distributions.
}
